requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
altair>=5.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.4
openai>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
scipy>=1.10.0
japanize-matplotlib>=1.1.3
python-dotenv
